digital light intensity display (design)
Design a digital light intensity display using Verilog.

The module should read light intensity data from a simulated sensor (take this value as input to the module) and display it on outputs. In this case, the relationship between sensor values and display values is:

Sensor Values (SV) range from 0 to 255, with 0 representing the lowest light intensity and 255 representing the highest light intensity.

Display Values (DV) are calculated non-linearly to create a more realistic response to light intensity. The relationship can be described as follows:

DV = (S * V ^ 2) / 255