digital temperature sensor (testbench in verilog)
Design a very basic digital temperature display using Verilog. The module reads temperature data from a simulated sensor (take this value as input to the module) and displays it on outputs in a simple format. The relationship between the sensor value and temperature is linear, where:

Sensor Value (SV) ranges from 0 to 255, where 0 represents the lowest temperature, and 255 represents the highest temperature.

Temperature (T) is calculated linearly as follows: T=SV^ * 2 - 20 degrees Celsius